backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
slowapi
fastapi-limiter
supabase